+++ sdk/mpr/writer_navigation.go
@@ -285,7 +285,7 @@ func buildCaptionBson(text string) bson.D {
 		{Key: "$ID", Value: idToBsonBinary(generateUUID())},
 		{Key: "$Type", Value: "Texts$Text"},
 		{Key: "Items", Value: bson.A{
-			int64(1),
+			int32(1),
 			bson.D{
 				{Key: "$ID", Value: idToBsonBinary(generateUUID())},
 				{Key: "$Type", Value: "Texts$Translation"},
